requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
groq==0.4.2
pandas==2.2.1
plotly==5.19.0 